Move files into classes/ folder and change imports to match
